Separate normalizeAnglePIPI into newly created lib/mathex.py
